# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Think Tank CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface SuccessEnvelope {
  success: boolean;
}

export interface Contact {
  id: string;
  contactId: string;
  fullName: string;
  /** @nullable */
  designation?: string | null;
  category: string;
  organization: string;
  sector: string;
  officialEmail: string;
  /** @nullable */
  personalEmail?: string | null;
  /** @nullable */
  phone?: string | null;
  city: string;
  state: string;
  relationshipLevel: string;
  sensitivityLevel: string;
  /** @nullable */
  lastVerifiedDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  confidentialNotes?: string | null;
  policyDomains?: string[];
  engagementScore?: number;
  archived: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateContactRequest {
  fullName: string;
  designation?: string;
  category: string;
  organization: string;
  sector: string;
  officialEmail: string;
  personalEmail?: string;
  phone?: string;
  city: string;
  state: string;
  relationshipLevel?: string;
  sensitivityLevel?: string;
  notes?: string;
  confidentialNotes?: string;
  policyDomains?: string[];
}

export interface UpdateContactRequest {
  fullName?: string;
  designation?: string;
  category?: string;
  organization?: string;
  sector?: string;
  officialEmail?: string;
  personalEmail?: string;
  phone?: string;
  city?: string;
  state?: string;
  relationshipLevel?: string;
  sensitivityLevel?: string;
  lastVerifiedDate?: string;
  notes?: string;
  confidentialNotes?: string;
  policyDomains?: string[];
  archived?: boolean;
}

export interface SubmitContactRequest {
  fullName: string;
  designation?: string;
  organization: string;
  officialEmail: string;
  phone?: string;
  city?: string;
  state?: string;
  notes?: string;
  policyDomains?: string[];
}

export interface Interaction {
  id: string;
  contactId: string;
  /** @nullable */
  contactName?: string | null;
  date: string;
  engagementType: string;
  policyDomain: string;
  outcome: string;
  followUpRequired: boolean;
  /** @nullable */
  followUpDate?: string | null;
  sensitivityFlag: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  createdBy?: string | null;
  createdAt: string;
}

export interface CreateInteractionRequest {
  contactId: string;
  date: string;
  engagementType: string;
  policyDomain: string;
  outcome: string;
  followUpRequired?: boolean;
  followUpDate?: string;
  sensitivityFlag?: boolean;
  notes?: string;
}

export interface UpdateInteractionRequest {
  date?: string;
  engagementType?: string;
  policyDomain?: string;
  outcome?: string;
  followUpRequired?: boolean;
  followUpDate?: string;
  sensitivityFlag?: boolean;
  notes?: string;
}

export interface CrmUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  role: string;
  status: string;
  createdAt: string;
}

export interface UpdateUserRoleRequest {
  role: string;
  status?: string;
}

export type DashboardStatsContactsByCategoryItem = {
  category: string;
  count: number;
};

export type DashboardStatsInteractionsByMonthItem = {
  month: string;
  count: number;
};

export type DashboardStatsRelationshipBreakdownItem = {
  level: string;
  count: number;
};

export type DashboardStatsRecentActivityItem = {
  id: string;
  action: string;
  timestamp: string;
  /** @nullable */
  user?: string | null;
};

export interface DashboardStats {
  totalContacts: number;
  totalInteractions: number;
  overdueFollowUps: number;
  contactsByCategory: DashboardStatsContactsByCategoryItem[];
  interactionsByMonth: DashboardStatsInteractionsByMonthItem[];
  relationshipBreakdown: DashboardStatsRelationshipBreakdownItem[];
  agingContacts?: number;
  recentActivity: DashboardStatsRecentActivityItem[];
}

export interface AuditLog {
  id: string;
  /** @nullable */
  userEmail: string | null;
  action: string;
  entityType: string;
  /** @nullable */
  entityId: string | null;
  /** @nullable */
  details?: string | null;
  timestamp: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListContactsParams = {
  search?: string;
  category?: string;
  sector?: string;
  relationship_level?: string;
  archived?: boolean;
};

export type ListContacts200 = {
  contacts: Contact[];
};

export type ListInteractionsParams = {
  contact_id?: string;
  follow_up_required?: boolean;
  overdue?: boolean;
};

export type ListInteractions200 = {
  interactions: Interaction[];
};

export type ListUsers200 = {
  users: CrmUser[];
};

export type ListAuditLogsParams = {
  limit?: number;
};

export type ListAuditLogs200 = {
  logs: AuditLog[];
};