# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Think Tank CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: sector
          in: query
          required: false
          schema:
            type: string
        - name: relationship_level
          in: query
          required: false
          schema:
            type: string
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                required: [contacts]
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contact"

    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Duplicate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      operationId: archiveContact
      tags: [contacts]
      summary: Archive (soft delete) a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contact archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /contacts/submit:
    post:
      operationId: submitContact
      tags: [contacts]
      summary: Public/internal contact submission form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContactRequest"
      responses:
        "201":
          description: Submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /interactions:
    get:
      operationId: listInteractions
      tags: [interactions]
      summary: List all interactions
      parameters:
        - name: contact_id
          in: query
          required: false
          schema:
            type: string
        - name: follow_up_required
          in: query
          required: false
          schema:
            type: boolean
        - name: overdue
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of interactions
          content:
            application/json:
              schema:
                type: object
                required: [interactions]
                properties:
                  interactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Interaction"

    post:
      operationId: createInteraction
      tags: [interactions]
      summary: Log a new interaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInteractionRequest"
      responses:
        "201":
          description: Interaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /interactions/{id}:
    get:
      operationId: getInteraction
      tags: [interactions]
      summary: Get an interaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interaction detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    put:
      operationId: updateInteraction
      tags: [interactions]
      summary: Update an interaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInteractionRequest"
      responses:
        "200":
          description: Updated interaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    Contact:
      type: object
      required: [id, contactId, fullName, category, organization, sector, officialEmail, city, state, relationshipLevel, sensitivityLevel, archived, createdAt, updatedAt]
      properties:
        id:
          type: string
        contactId:
          type: string
        fullName:
          type: string
        designation:
          type: [string, 'null']
        category:
          type: string
        organization:
          type: string
        sector:
          type: string
        officialEmail:
          type: string
        personalEmail:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        city:
          type: string
        state:
          type: string
        relationshipLevel:
          type: string
        sensitivityLevel:
          type: string
        lastVerifiedDate:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        confidentialNotes:
          type: [string, 'null']
        policyDomains:
          type: array
          items:
            type: string
        engagementScore:
          type: number
        archived:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    CreateContactRequest:
      type: object
      required: [fullName, category, organization, sector, officialEmail, city, state]
      properties:
        fullName:
          type: string
        designation:
          type: string
        category:
          type: string
        organization:
          type: string
        sector:
          type: string
        officialEmail:
          type: string
          format: email
        personalEmail:
          type: string
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        relationshipLevel:
          type: string
        sensitivityLevel:
          type: string
        notes:
          type: string
        confidentialNotes:
          type: string
        policyDomains:
          type: array
          items:
            type: string
    UpdateContactRequest:
      type: object
      properties:
        fullName:
          type: string
        designation:
          type: string
        category:
          type: string
        organization:
          type: string
        sector:
          type: string
        officialEmail:
          type: string
        personalEmail:
          type: string
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        relationshipLevel:
          type: string
        sensitivityLevel:
          type: string
        lastVerifiedDate:
          type: string
        notes:
          type: string
        confidentialNotes:
          type: string
        policyDomains:
          type: array
          items:
            type: string
        archived:
          type: boolean
    SubmitContactRequest:
      type: object
      required: [fullName, organization, officialEmail]
      properties:
        fullName:
          type: string
        designation:
          type: string
        organization:
          type: string
        officialEmail:
          type: string
          format: email
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        notes:
          type: string
        policyDomains:
          type: array
          items:
            type: string
    Interaction:
      type: object
      required: [id, contactId, date, engagementType, policyDomain, outcome, followUpRequired, sensitivityFlag, createdAt]
      properties:
        id:
          type: string
        contactId:
          type: string
        contactName:
          type: [string, 'null']
        date:
          type: string
        engagementType:
          type: string
        policyDomain:
          type: string
        outcome:
          type: string
        followUpRequired:
          type: boolean
        followUpDate:
          type: [string, 'null']
        sensitivityFlag:
          type: boolean
        notes:
          type: [string, 'null']
        createdBy:
          type: [string, 'null']
        createdAt:
          type: string
    CreateInteractionRequest:
      type: object
      required: [contactId, date, engagementType, policyDomain, outcome]
      properties:
        contactId:
          type: string
        date:
          type: string
        engagementType:
          type: string
        policyDomain:
          type: string
        outcome:
          type: string
        followUpRequired:
          type: boolean
        followUpDate:
          type: string
        sensitivityFlag:
          type: boolean
        notes:
          type: string
    UpdateInteractionRequest:
      type: object
      properties:
        date:
          type: string
        engagementType:
          type: string
        policyDomain:
          type: string
        outcome:
          type: string
        followUpRequired:
          type: boolean
        followUpDate:
          type: string
        sensitivityFlag:
          type: boolean
        notes:
          type: string
...