# lib/api-zod/src/generated/types/createContactRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Think Tank CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactRequest {
  fullName: string;
  designation?: string;
  category: string;
  organization: string;
  sector: string;
  officialEmail: string;
  personalEmail?: string;
  phone?: string;
  city: string;
  state: string;
  relationshipLevel?: string;
  sensitivityLevel?: string;
  notes?: string;
  confidentialNotes?: string;
  policyDomains?: string[];
}